Added getBound, collision and gmeOver functions
Now getBounds returns a rectangle object that surrounds the actual
object. With that we can use the given interection function in
rectangle object that returns true if any collition occurs. We use it in
collision function of the ball object to change the direction of the
ball when it hits the racket.